+++ test/decoding.js
@@ -13,7 +13,7 @@ const Branch = require('../src/models/branch');
 const WorkingState = require('../src/models/workingState');
 const RepositoryState = require('../src/models/repositoryState');
 
-describe('Decoding, encoding', function() {
+describe('Decoding, encoding', () => {
 
     const blob = Blob.createFromString('Test');
 
@@ -76,7 +76,7 @@ describe('Decoding, encoding', function() {
         };
     }
 
-    it('should encode and decode back a Blob', function() {
+    it('should encode and decode back a Blob', () => {
         const encoded = Blob.encode(blob);
         const decoded = Blob.decode(encoded);