+++ HISTORY.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### 7.9.0
+
+- Added `CommitUtils.fetchComparison` method
+- Added models for `FileDiff` and `Comparison`
+
 ### 7.8.0
 
 - Added `RepoUtils.isClean` method